+++ src/main/java/com/nocaffeine/ssgclone/SsgCloneApplication.java
@@ -6,7 +6,10 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@OpenAPIDefinition(servers = {@Server(url = "https://nocaffein.shop", description = "노카페인")})
+@OpenAPIDefinition(servers = {
+		@Server(url = "https://nocaffein.shop", description = "노카페인"),
+		@Server(url = "http://localhost:8080", description = "로컬")
+})
 @EnableJpaAuditing
 @SpringBootApplication
 public class SsgCloneApplication {
