+++ src/main/java/com/nocaffeine/ssgclone/member/application/AuthService.java
@@ -1,5 +1,7 @@
 package com.nocaffeine.ssgclone.member.application;
 
+import com.nocaffeine.ssgclone.member.dto.request.AuthCheckRequestDto;
+import com.nocaffeine.ssgclone.member.dto.request.AuthEmailRequestDto;
 import com.nocaffeine.ssgclone.member.dto.request.SnsMemberAddRequestDto;
 import com.nocaffeine.ssgclone.member.dto.request.SnsMemberLoginRequestDto;
 import com.nocaffeine.ssgclone.member.dto.response.TokenResponseDto;
@@ -7,4 +9,9 @@ import com.nocaffeine.ssgclone.member.dto.response.TokenResponseDto;
 public interface AuthService {
     void snsAddMember(SnsMemberAddRequestDto snsMemberAddRequestDto);
     TokenResponseDto snsLogin(SnsMemberLoginRequestDto snsMemberLoginRequestDto);
+
+
+    void emailAuth(AuthEmailRequestDto authEmailRequestDto);
+
+    void emailAuthCodeCheck(AuthCheckRequestDto authCheckRequestDto);
 }
