+++ src/main/java/com/nocaffeine/ssgclone/member/application/MemberService.java
@@ -18,8 +18,4 @@ public interface MemberService {
 
     void removeMember(String memberUuid);
 
-    void emailAuth(AuthEmailRequestDto authEmailRequestDto);
-
-    void emailAuthCodeCheck(AuthCheckRequestDto authCheckRequestDto);
-
 }
