+++ src/main/java/com/nocaffeine/ssgclone/member/application/MemberServiceImp.java
@@ -1,13 +1,10 @@
 package com.nocaffeine.ssgclone.member.application;
 
-import com.nocaffeine.ssgclone.common.EmailProvider;
 import com.nocaffeine.ssgclone.common.exception.BaseException;
-import com.nocaffeine.ssgclone.member.domain.EmailAuth;
 import com.nocaffeine.ssgclone.member.domain.Member;
 import com.nocaffeine.ssgclone.member.dto.request.*;
 import com.nocaffeine.ssgclone.member.dto.response.MemberDetailResponseDto;
 import com.nocaffeine.ssgclone.member.dto.response.TokenResponseDto;
-import com.nocaffeine.ssgclone.member.infrastructure.EmailAuthRepository;
 import com.nocaffeine.ssgclone.member.infrastructure.MemberRepository;
 import com.nocaffeine.ssgclone.common.security.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
@@ -17,8 +14,6 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
-import java.util.Optional;
 import java.util.UUID;
 
 import static com.nocaffeine.ssgclone.common.exception.BaseResponseStatus.*;
@@ -33,8 +28,6 @@ public class MemberServiceImp implements MemberService {
     private final MemberRepository memberRepository;
     private final AuthenticationManager authenticateManager;
     private final JwtTokenProvider jwtTokenProvider;
-    private final EmailProvider emailProvider;
-    private final EmailAuthRepository emailAuthRepository;
 
 
     /**
@@ -164,69 +157,5 @@ public class MemberServiceImp implements MemberService {
         memberRepository.delete(member);
     }
 
-
-    /**
-     * 이메일 인증코드 발송
-     */
-    @Override
-    @Transactional
-    public void emailAuth(AuthEmailRequestDto authEmailRequestDto) {
-        duplicationEmail(authEmailRequestDto.getEmail());
-
-        String authCode = createAuthCode();
-
-        boolean isSuccess = emailProvider.sendAuthMail(authEmailRequestDto.getEmail(), authCode);
-
-        if(!isSuccess){
-            throw new BaseException(MASSAGE_SEND_FAILED);
-        }
-
-        log.info("이메일 인증코드 : {}", authCode);
-
-        Optional<EmailAuth> email = emailAuthRepository.findByEmail(authEmailRequestDto.getEmail());
-
-        if(email.isPresent()){
-            email.get().updateAuthCode(authCode);
-        } else {
-            EmailAuth emailAuth = EmailAuth.builder()
-                    .email(authEmailRequestDto.getEmail())
-                    .authCode(authCode)
-                    .expireDate(LocalDateTime.now())
-                    .build();
-
-            emailAuthRepository.save(emailAuth);
-        }
-    }
-
-    public static String createAuthCode() {
-        String authCode = "";
-
-        for(int count = 0; count < 6; count++){
-            authCode += (int)(Math.random() * 10);
-        }
-
-        return authCode;
-    }
-
-    /**
-     * 이메일 인증코드 확인
-     */
-    @Override
-    @Transactional
-    public void emailAuthCodeCheck(AuthCheckRequestDto authCheckRequestDto) {
-        EmailAuth emailAuth = emailAuthRepository.findByEmail(authCheckRequestDto.getEmail())
-                .orElseThrow(() -> new BaseException(NO_EXIST_AUTH));
-
-        if(!emailAuth.getAuthCode().equals(authCheckRequestDto.getAuthCode())){
-            throw new BaseException(MASSAGE_VALID_FAILED);
-        }
-
-        if (emailAuth.getExpireDate().isBefore(LocalDateTime.now())) {
-            // 인증 코드가 만료된 경우
-            throw new BaseException(EXPIRED_AUTH_CODE);
-        }
-
-        emailAuthRepository.delete(emailAuth);
-    }
 }
