+++ src/main/java/com/nocaffeine/ssgclone/member/presentation/MemberController.java
@@ -24,21 +24,21 @@ public class MemberController {
     private final MemberService memberService;
     private final JwtTokenProvider jwtTokenProvider;
 
-    @Operation(summary = "이메일 중복 검증", description = "이메일 중복 검증", tags = {"Duplication Email"})
+    @Operation(summary = "이메일 중복 검증", description = "이메일 중복 검증", tags = {"Member"})
     @GetMapping("/duplication")
     public CommonResponse<String> duplicationEmail(@RequestParam String email) {
         memberService.duplicationEmail(email);
         return CommonResponse.success("이메일 중복 검증 성공");
     }
 
-    @Operation(summary = "회원가입", description = "회원가입", tags = {"Sign Up"})
+    @Operation(summary = "회원가입", description = "회원가입", tags = {"Member"})
     @PostMapping("/member")
     public CommonResponse<String> memberCreate(@RequestBody MemberSaveRequestVo memberSaveRequestVo) {
         memberService.addMember(MemberSaveRequestDto.voToDto(memberSaveRequestVo));
         return CommonResponse.success("회원가입 성공");
     }
 
-    @Operation(summary = "로그인", description = "로그인", tags = {"Log In"})
+    @Operation(summary = "로그인", description = "로그인", tags = {"Member"})
     @PostMapping("/member/login")
     public ResponseEntity<CommonResponse<TokenResponseDto>> logIn(@RequestBody MemberLoginRequestVo memberLoginRequestVo) {
         TokenResponseDto tokenResponseDto = memberService.logIn(MemberLoginRequestDto.voToDto(memberLoginRequestVo));
@@ -47,7 +47,7 @@ public class MemberController {
                 .body(CommonResponse.success("로그인 성공"));
     }
 
-    @Operation(summary = "비밀번호 변경", description = "비밀번호 변경", tags = {"Change Password"})
+    @Operation(summary = "비밀번호 변경", description = "비밀번호 변경", tags = {"Member"})
     @PutMapping("/member/password")
     public CommonResponse<String> changePassword(@RequestBody MemberPasswordRequestVo memberPasswordRequestVo) {
         String memberUuid = jwtTokenProvider.validateAndGetUserUuid(jwtTokenProvider.getHeader());
@@ -55,34 +55,19 @@ public class MemberController {
         return CommonResponse.success("비밀번호 변경 성공");
     }
 
-    @Operation(summary = "회원 상세 정보", description = "회원 상세 정보", tags = {"Member Detail"})
+    @Operation(summary = "회원 상세 정보", description = "회원 상세 정보", tags = {"Member"})
     @GetMapping("/member")
     public CommonResponse<MemberDetailResponseVo> memberDetail() {
         String memberUuid = jwtTokenProvider.validateAndGetUserUuid(jwtTokenProvider.getHeader());
         return CommonResponse.success("회원 상세 조회 성공",MemberDetailResponseDto.dtoToVo(memberService.findMember(memberUuid)));
     }
 
-    @Operation(summary = "회원 탈퇴", description = "회원 탈퇴", tags = {"Remove Member"})
+    @Operation(summary = "회원 탈퇴", description = "회원 탈퇴", tags = {"Member"})
     @DeleteMapping("/member")
     public CommonResponse<String> memberRemove() {
         String memberUuid = jwtTokenProvider.validateAndGetUserUuid(jwtTokenProvider.getHeader());
         memberService.removeMember(memberUuid);
         return CommonResponse.success("회원 탈퇴 성공");
     }
-
-    @Operation(summary = "이메일 인증코드 발송", description = "이메일 인증코드 발송", tags = {"Email Certification"})
-    @PostMapping("/email-auth")
-    public CommonResponse<String> emailAuth(@RequestBody AuthEmailRequestVo authEmailRequestVo) {
-        memberService.emailAuth(AuthEmailRequestDto.voToDto(authEmailRequestVo));
-        return CommonResponse.success("이메일 인증코드 발송 성공");
-    }
-
-    @Operation(summary = "이메일 인증코드 확인", description = "이메일 인증코드 확인", tags = {"Email Certification"})
-    @PostMapping("/email-auth/check")
-    public CommonResponse<String> emailAuthCodeCheck(@RequestBody AuthCheckRequestVo authCheckRequestVo) {
-        memberService.emailAuthCodeCheck(AuthCheckRequestDto.voToDto(authCheckRequestVo));
-        return CommonResponse.success("이메일 인증코드 확인 성공");
-    }
-
 
 }
